+++ SourceCode/Database/dbo/Views/ModuleConsumerWaybill.sql
@@ -49,6 +49,7 @@ SELECT
 	RECEIVER.IsModuleStation AS DestinationIsModuleStation,
 	RECEIVER.SpecialCargoName,
 	RECEIVER.SpecificWagonClass,
+	RECEIVER.QuantityIsBearer,
 	RECEIVER.QuantityUnitId,
 	RECEIVER.Quantity,
 	RECEIVER.QuantityUnitResourceCode AS QuanityUnitResourceName,